% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch_boxplot.R
\name{notch_boxplot}
\alias{notch_boxplot}
\title{Title: Dual-notched boxplot using ggplot2}
\usage{
notch_boxplot(
  data,
  group_col,
  value_col,
  show_mean_ci = TRUE,
  show_med_ci = TRUE,
  width = 0.4,
  indent_pct = 0.125,
  mean_side = "left",
  med_side = "right",
  line_color = "black",
  fill_color = "white",
  mean_color = "black",
  outlier_size = 2
)
}
\arguments{
\item{data}{The data frame containing the data.}

\item{group_col}{The column name for grouping data (character string).}

\item{value_col}{The column name for the values to plot (character string).}

\item{show_mean_ci}{Logical. Whether to display the mean notch. Default is TRUE.}

\item{show_med_ci}{Logical. Whether to display the median notch. Default is TRUE.}

\item{width}{Numeric. The width of the box. Default is 0.4.}

\item{indent_pct}{Numeric. The depth of the notch indentation as a percentage of the width. Default is 0.125.}

\item{mean_side}{Character. The side of the box to display the mean notch ("left" or "right"). Default is "left".}

\item{med_side}{Character. The side of the box to display the median notch ("left" or "right"). Default is "right".}

\item{line_color}{Character. The color of the box borders, whiskers, and median line. Default is "black".}

\item{fill_color}{Character. The default fill color for the box. Default is "white".}

\item{mean_color}{Character. The color of the dashed mean line. Default is "black".}

\item{outlier_size}{Numeric. The size of the points representing outliers. Default is 2.}
}
\value{
A ggplot object representing the dual-notched boxplot.
}
\description{
This function generates a notched boxplot, which can simultaneously compare both the group means and medians.
For theoretical details, please refer to Kong et al. (2026).
}
\examples{
# Example using the built-in ToothGrowth dataset
notch_boxplot(data = ToothGrowth,
              group_col = "supp",
              value_col = "len",
              show_mean_ci = TRUE,
              show_med_ci = FALSE)

# Example using the built-in iris dataset
notch_boxplot(data = iris,
              group_col = "Species",
              value_col = "Sepal.Width")
}
\references{
Deru Kong, Xuming He, WenWu Wang, Tiejun Tong (2026). Dual-Notched Boxplot: A New Visualization for the Simultaneous Comparison of Means and Medians.
}
